Chore: Migrate mime-utils to TypeScript
