+++ packages/app-desktop/gui/NoteEditor/utils/resourceHandling.ts
@@ -15,7 +15,7 @@ import markupRenderOptions from './markupRenderOptions';
 import { fileExtension, filename, safeFileExtension, safeFilename } from '@joplin/utils/path';
 const joplinRendererUtils = require('@joplin/renderer').utils;
 const { clipboard } = require('electron');
-const mimeUtils = require('@joplin/lib/mime-utils.js').mime;
+import * as mimeUtils from '@joplin/lib/mime-utils';
 const md5 = require('md5');
 const path = require('path');
