+++ packages/app-mobile/utils/shim-init-react.js
@@ -7,7 +7,7 @@ const FsDriverRN = require('./fs-driver/fs-driver-rn').default;
 const { Buffer } = require('buffer');
 const { Linking, Platform } = require('react-native');
 const showMessageBox = require('./showMessageBox.js').default;
-const mimeUtils = require('@joplin/lib/mime-utils.js').mime;
+const mimeUtils = require('@joplin/lib/mime-utils.js');
 const { basename, fileExtension } = require('@joplin/lib/path-utils');
 const uuid = require('@joplin/lib/uuid').default;
 const Resource = require('@joplin/lib/models/Resource').default;
