+++ packages/lib/import-enex.ts
@@ -16,7 +16,7 @@ const { enexXmlToHtml } = require('./import-enex-html-gen.js');
 const md5 = require('md5');
 const { Base64Decode } = require('base64-stream');
 const md5File = require('md5-file');
-const { mime } = require('./mime-utils');
+import * as mime from './mime-utils';
 
 // const Promise = require('promise');
 const fs = require('fs-extra');
