+++ packages/tools/fetchPatreonPosts.js
@@ -7,7 +7,7 @@ const { patreonOauthToken } = require('./tool-utils');
 const HtmlToMd = require('@joplin/lib/HtmlToMd').default;
 const { dirname, filename, basename } = require('@joplin/lib/path-utils');
 const markdownUtils = require('@joplin/lib/markdownUtils').default;
-const mimeUtils = require('@joplin/lib/mime-utils.js').mime;
+const mimeUtils = require('@joplin/lib/mime-utils.js');
 const { mimeTypeFromHeaders } = require('@joplin/lib/net-utils');
 const shim = require('@joplin/lib/shim').default;
 const moment = require('moment');
